python/requirements.txt
pyserial
pyautogui
pydirectinput